requirements.txt
cffi==1.15.1
cryptography==38.0.1
jeepney==0.8.0
pycparser==2.21
PyYAML==6.0
SecretStorage==3.3.3
